apply: Add option to remove or pick patches

+++ apply.sh
@@ -10,14 +10,33 @@ rootdirectory="$PWD"
 
 dirs="frameworks/base packages/apps packages/modules lineage-sdk system/core"
 
-for dir in $dirs ; do
+echo -e "\n1.Remove all patches\n2.Pick all patches"
+read -p "Your choice: " num
+case $num in 
+    1|2|3|4)
+esac
+
+if [ $num = '1' ]; then
+    for dir in $dirs ; do
 	cd $rootdirectory
 	cd $dir
 	echo "Cleaning $dir patches..."
 	git reset --hard
 	git clean -f -d
+done
+
+cd $rootdirectory
+    
+elif [ $num = '2' ]; then
+    for dir in $dirs ; do
+	cd $rootdirectory
+	cd $dir
 	echo "Applying $dir patches..."
 	git apply $rootdirectory/patches/$dir/*.patch
 done
 
 cd $rootdirectory
+    
+else echo "Invalid input, aborting!"
+
+fi
